chart_tool/requirements.txt
dash>=2.0
plotly>=5.0
pandas>=1.3
python-dotenv
SQLAlchemy>=1.4
pymysql
pytest
requests>=2.0
fastapi>=0.70
uvicorn[standard]
